yad example of sysinfo
A shell script with a yad wrapper for a front end gui. I don't recall where I found this example but it has helped me traverse building yad interfaces a lot more than the documentation that is out there. This was a great learn by example tool. I hope it helps others.